requirements.txt
pdfplumber
faiss-cpu
langchain
sentence-transformers
langchain-text-splitters